requirements.txt
tensorflow-gpu==1.8.0
tensorboard==1.8.0
numpy
